[Python] Choose to add or not date prefix if there is some already
Sometimes file have date prefix, but it is wrong (or exif lies). In this
case user can choose if he wants to add correct prefix ot not.

This change is kinda weak, 'cause it is difficult to define if there is
a date, or because of adding new prefix, not replacing existing one.
Can't decide if it should be replaced. Leave it as is for now.